admin/mute: don't crash if an admin tries to mute another admin
this actually is supported by the bot, it'll delete the muted admin's
messages if the person who muted has message nuking rights. the
exception is caught in the function that deletes said messages but the
bot crashes when trying to edit another admin's permissions.

let's catch that exception and continue normally.

Change-Id: I7a8eb0c6ccf832b4c84bc27fecc3487c607a7ee4
Signed-off-by: KenHV <kenhv@null.net>